:truck: [#344] Moved interface file
